+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ACCOUNTING REPORT : display popup message on click of 'exportBtn'.
 ## Bug Fixes
 - EBICS USER : fix for strange import - log
+- LEAVE REQUEST: Updated calendar filter
 
 ## [5.0.11] - 2019-12-19
 ## Improvements